requirements.txt
fastapi
uvicorn
python-multipart
openai
chromadb
PyMuPDF
docx2txt
requests
python-dotenv
